Refactor robot control structure and update import paths
- Created a new `robot_control.py` module within the `utils/robot` directory to encapsulate robot control functions, including remote control commands and status checks.
- Updated import paths across various modules to reflect the new structure, ensuring consistent access to robot control functions.
- Introduced new enums in `robot_enums.py` for better organization of robot modes, safety statuses, and program states.
- Enhanced the Python import helper with updated paths for improved code completion related to robot functionalities.
